Fixes to goToHour and goToDate().
Allow goToDate to be called before the weekview is drawn for the first time.

goToDate relies on mWidthPerDay, which itself relies on View#getWidth() which can't be called during initialisation. Therefore the goToDate variable is cached until the first time the weekview is drawn.

Fixes #45:
1. Date is cloned before use (in goToDate)
2. goToHour doesn't throw errors anymore

Fixes #55:
goToDate now uses an animation

Fixes #65:
If a dimension is altered and is followed by a goToDate or goToHour, then the hour or date to go to, is cached.