Add WeekViewDisplayable interface
Instead of having to convert one's own data classes to WeekViewEvent, the WeekViewDisplayable interface includes a toWeekViewEvent() method. The conversion can be handled in the data class and does not add boilerplate in your controlling classes.

As a result, onMonthChanged() can return a list of objects that implement WeekViewDisplayable and does not require a list of WeekViewEvents.