Add workaround to prevent glitch when drawing day background and grid
in StaticActivity, when navigating from a week with all-day events to a week without them, the area previously used to show all-day events was not used to draw the normal events. By posting an invalidate, this problem appears to be fixed for now. I will look into how to prevent it in the first place.