Simplify events fetching (#102)
* Make minor code tweaks

* Add initial implementation of AsyncLoader

The idea is that AsyncLoader allows for a cleaner usage of WeekView when using a reactive architecture. The sample app's AsyncActivity shows an example using a ViewModel that fetches events from an API. Instead of having to provide an OnMonthChangeListener and using WeekView.notifyDataSetChanged(), it simply calls WeekView.submit() when the events are fetched.

If the ViewModel should be notified whenever WeekView requires more events (e.g. due to the user scrolling), WeekView offers a onLoadMore() callback to do so. It operates like an OnMonthChangeListener, but doesn't return anything.

* Only invalidate view if events in visible date range changed

* Consider case where submit() submits an empty list

* Refactor EventCache and EventChipsProvider

* Replace MonthLoader with EventsLoader, adapt class names

* Refactor isRequired() for AllDayEventsUpdater and HeaderRowHeightUpdater

* Add OnLoadMoreListener interface

* Fix issue where events loading behaved incorrectly

* Fix issue where event chip cache wasn't cleared properly

* Update behavior when calling WeekView.submit()

If you submit a new list of events, the existing events will be cleared before the new ones will be added.

* Make minor code tweak

* Make minor performance improvements

* Fix broken tests

* Fix incorrect behavior in AsyncLoader