Add support for showing blocked times (#180)
* Add support for showing blocked times

* Rename location to subtitle and allow it for WeekViewEntity.BlockedTime

* Remove generic from WeekViewDisplayable

The generic is now only used for WeekViewEntity.Event. This avoids having to provide a generic type for WeekViewEntity.BlockedTime

* Deprecate WeekViewDisplayable and move WeekViewEntity creation to adapter

This commit deprecates WeekViewDisplayable. Instead of requiring data classes to implement this interface for constructing a WeekViewEntity (or the previous WeekViewEvent), this construction now occurs in the adapter’s onCreateEntity() method.

This enables a better separation of concerns and makes it possible to pass a Context to the construction site, allowing you to resolve resources directly. This can be helpful when deciding on which color to use depending on the device’s current theme.

* Make minor code tweaks

* Fix incorrect drawing of dotted background pattern

* Prevent duplicate work when creating event chips from submitted entities

* Tweak API for adding a Pattern to a WeekViewEntity.Style

* Add option to set the direction of Pattern.Lined

* Make sure blocked times are drawn correctly in RTL